Introduce kernel rebuild CI test
Introduce baseos-qe.koji-build.scratch-build.validation as a new
gating test for annobin. The test makes sure that the annobin
update under test doesn't break the kernel build process. The
testing kernel build happens in koji in a side-tag buildroot
containing the annobin build under test.

The source code of this new testcase lives here:
https://github.com/fedora-ci/scratch-build-test

In case of a failure that is considered a false positive, the test
result can be waived using Bodhi this way:
https://docs.fedoraproject.org/en-US/ci/gating/#_waive